handle when there is an uncreated chapter
If the file _data/contents.yml had reference to a chapter which does not
exist for the default languge, this code would error. Commenting out the
three which did not exist fixed the problem.

at time of writing, these were:
  - types
  - algebraic-data-types
  - lists-tuples
  - pattern-matching-control-flow

instead fo commenting those out though, I decided to make the code a
little more robust to handle this case.